Render/requirements.txt
Flask==3.1.1
peewee==3.18.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
gunicorn==23.0.0
Flask-Cors==4.0.0